students_qr/requirements.txt
asgiref         3.10.0
Django          5.2.7
pillow          12.0.0
pip             24.0
psycopg2-binary 2.9.11
python-decouple 3.8
qrcode          8.2
sqlparse        0.5.3